fix: add error handlings

+++ executor/wal.go
@@ -11,10 +11,11 @@ import (
 	"sync"
 	"time"
 
+	"go.uber.org/zap"
+
 	"github.com/alpacahq/marketstore/v4/executor/buffile"
 
 	"github.com/alpacahq/marketstore/v4/executor/wal"
-	"github.com/alpacahq/marketstore/v4/plugins/trigger"
 	"github.com/alpacahq/marketstore/v4/utils/io"
 	"github.com/alpacahq/marketstore/v4/utils/log"
 )
@@ -75,7 +76,9 @@ func NewWALFile(rootDir string, owningInstanceID int64, rs ReplicationSender,
 		log.Error("%v: Can not create new WALFile - Error: %v", io.GetCallerFileContext(0), err)
 		return nil, fmt.Errorf("can not create new WALFile: %w", err)
 	}
-	wf.WriteStatus(wal.OPEN, wal.NOTREPLAYED)
+	if err := wf.WriteStatus(wal.OPEN, wal.NOTREPLAYED); err != nil {
+		return nil, fmt.Errorf("failed to write NOT_REPLAYED status to wal: %w", err)
+	}
 
 	return wf, nil
 }
@@ -102,7 +105,9 @@ func TakeOverWALFile(filePath string) (wf *WALFileType, err error) {
 	wf.OwningInstanceID = owningInstanceID
 
 	// We call this to take over the file by writing our PID to it
-	wf.WriteStatus(fileStatus, replayState)
+	if err := wf.WriteStatus(fileStatus, replayState); err != nil {
+		return nil, fmt.Errorf("failed to write replayState:%v to walfile: %w", replayState, err)
+	}
 
 	return wf, nil
 }
@@ -131,9 +136,11 @@ func (wf *WALFileType) open(filePath string) error {
 	return nil
 }
 
-func (wf *WALFileType) close(replayStatus wal.ReplayStateEnum) {
-	wf.WriteStatus(wal.CLOSED, replayStatus)
-	wf.FilePtr.Close()
+func (wf *WALFileType) close(replayStatus wal.ReplayStateEnum) error {
+	if err := wf.WriteStatus(wal.CLOSED, replayStatus); err != nil {
+		return fmt.Errorf("failed to write CLOSED status to wal: %w", err)
+	}
+	return wf.FilePtr.Close()
 }
 
 func (wf *WALFileType) Delete(callersInstanceID int64) (err error) {
@@ -161,7 +168,9 @@ func (wf *WALFileType) Delete(callersInstanceID int64) (err error) {
 		return errors.New("WAL File needs replay, can not delete")
 	}
 
-	wf.close(wal.REPLAYED)
+	if err2 := wf.close(wal.REPLAYED); err2 != nil {
+		return fmt.Errorf("failed to close walfile: %w", err2)
+	}
 	if err = os.Remove(wf.FilePtr.Name()); err != nil {
 		log.Error(io.GetCallerFileContext(0) + ": Can not remove WALFile")
 		return fmt.Errorf("cannot remove WALFile %s: %w", wf.FilePtr.Name(), err)
@@ -203,14 +212,11 @@ func (wf *WALFileType) QueueWriteCommand(wc *wal.WriteCommand) {
 }
 
 // FlushToWAL A.k.a. Commit transaction.
+//	Here we flush the contents of the write cache to:
+//	- Primary storage via the OS write cache - data is visible to readers
+//	- WAL file with synchronization to physical storage - in case we need to recover from a crash
 func (wf *WALFileType) FlushToWAL() (err error) {
 	// walBypass = true // Bypass all writing to the WAL File, leaving the writes to the primary
-
-	/*
-		Here we flush the contents of the write cache to:
-		- Primary storage via the OS write cache - data is visible to readers
-		- WAL file with synchronization to physical storage - in case we need to recover from a crash
-	*/
 
 	// Count of WT Sets in this TG as of now
 	if wf.txnPipe == nil {
@@ -225,14 +231,16 @@ func (wf *WALFileType) FlushToWAL() (err error) {
 	}
 
 	if !wf.walBypass {
-		canWrite, err := wf.CanWrite("WriteTG", wf.OwningInstanceID)
+		canWrite, err := wf.CanWrite(wf.OwningInstanceID)
 		// TODO: error handling to move walFile to a temporary file and create a new one when walFile is corrupted.
 		if err != nil || !canWrite {
 			panic("Failed attempt to write to WAL")
 		}
 
 		// WAL Transaction Preparing Message
-		wf.WriteTransactionInfo(wf.txnPipe.TGID(), WAL, PREPARING)
+		if err := wf.WriteTransactionInfo(wf.txnPipe.TGID(), WAL, PREPARING); err != nil {
+			return fmt.Errorf("failed to write TRANSACTION PREPARING message to wal: %w", err)
+		}
 	}
 
 	// Serialize all data to be written except for the size of this buffer
@@ -270,20 +278,32 @@ func (wf *WALFileType) FlushCommandsToWAL(writeCommands []*wal.WriteCommand) (er
 		hash.Write(tgLenSerialized)
 		hash.Write(tgSerialized)
 
-		wf.FilePtr.Write(wf.initMessage(TGDATA)) // Write the Message ID to identify TG Data
+		// Write the Message ID to identify TG Data
+		if _, err := wf.FilePtr.Write(wf.initMessage(TGDATA)); err != nil {
+			return fmt.Errorf("failed to write Message ID to walfile: %w", err)
+		}
 		// Write the TG Data and the checksum and Sync()
-		wf.FilePtr.Write(tgLenSerialized)
-		wf.FilePtr.Write(tgSerialized)
+		if _, err := wf.FilePtr.Write(tgLenSerialized); err != nil {
+			return fmt.Errorf("failed to write TransactionGroup length to walfile: %w", err)
+		}
+		if _, err := wf.FilePtr.Write(tgSerialized); err != nil {
+			return fmt.Errorf("failed to write TransactionGroup data to walfile: %w", err)
+		}
 		cksum := hash.Sum(nil)
-		wf.FilePtr.Write(cksum) // Checksum
-
+		if _, err := wf.FilePtr.Write(cksum); err != nil { // Checksum
+			return fmt.Errorf("failed to write TransactionGroup checksum to walfile: %w", err)
+		}
 		// WAL Transaction Commit Complete Message
 		TGID := wf.txnPipe.TGID()
-		wf.WriteTransactionInfo(TGID, WAL, COMMITCOMPLETE)
+		if err := wf.WriteTransactionInfo(TGID, WAL, COMMITCOMPLETE); err != nil {
+			return fmt.Errorf("failed to write COMMITCOMPLETE status to walfile: %w", err)
+		}
 		wf.lastCommittedTGID = TGID
 		wf.txnPipe.IncrementTGID()
 
-		wf.FilePtr.Sync() // Flush the OS buffer
+		if err := wf.FilePtr.Sync(); err != nil { // Flush the OS buffer
+			return fmt.Errorf("failed to flush wal data: %w", err)
+		}
 
 		// send transaction to replicas
 		if wf.ReplicationSender != nil {
@@ -302,7 +322,7 @@ func (wf *WALFileType) FlushCommandsToWAL(writeCommands []*wal.WriteCommand) (er
 			log.Error(fmt.Sprintf("failed to write data to file %s: %s", keyPath, err.Error()))
 		}
 		for i, buffer := range writes {
-			wf.tpd.AppendRecord(keyPath, trigger.Record(buffer.IndexAndPayload()))
+			wf.tpd.AppendRecord(keyPath, buffer.IndexAndPayload())
 			writes[i] = nil // for GC
 		}
 		writesPerFile[keyPath] = nil // for GC
@@ -369,7 +389,11 @@ func writeFixedBuffer(writes []wal.OffsetIndexBuffer, fullPath string) error {
 		log.Error("cannot open file %s for write transaction commit: %v", fullPath, err)
 		return err
 	}
-	defer fp.Close()
+	defer func() {
+		if err2 := fp.Close(); err2 != nil {
+			log.Error("failed to close walfile", zap.Error(err2))
+		}
+	}()
 
 	for _, buffer := range writes {
 		if err = WriteBufferToFile(fp, buffer); err != nil {
@@ -387,7 +411,11 @@ func writeVariableLengthBuffer(writes []wal.OffsetIndexBuffer, fullPath string, 
 		log.Error("cannot open file %s for write transaction commit: %v", fullPath, err)
 		return err
 	}
-	defer fp.Close()
+	defer func() {
+		if err2 := fp.Close(); err2 != nil {
+			log.Error("failed to close walfile", zap.Error(err2))
+		}
+	}()
 
 	for _, buffer := range writes {
 		if err = WriteBufferToFileIndirect(fp, buffer, varRecLen); err != nil {
@@ -429,15 +457,22 @@ func (wf *WALFileType) CreateCheckpoint() error {
 	}
 	if wf.walBypass {
 		io.Syncfs()
-	} else {
-		// WAL Transaction Preparing Message
-		// Get the latest TGID and write a prepare message
-		TGID := wf.lastCommittedTGID
-		wf.WriteTransactionInfo(TGID, CHECKPOINT, PREPARING)
-		// Sync the filesystem, after this point the filesystem cache data is committed to disk
-		io.Syncfs()
-		wf.WriteTransactionInfo(TGID, CHECKPOINT, COMMITCOMPLETE)
+		wf.lastCommittedTGID = 0
+		return nil
 	}
+
+	// WAL Transaction Preparing Message
+	// Get the latest TGID and write a prepare message
+	TGID := wf.lastCommittedTGID
+	if err := wf.WriteTransactionInfo(TGID, CHECKPOINT, PREPARING); err != nil {
+		return fmt.Errorf("write PREPARING transaction info: %w", err)
+	}
+	// Sync the filesystem, after this point the filesystem cache data is committed to disk
+	io.Syncfs()
+	if err := wf.WriteTransactionInfo(TGID, CHECKPOINT, COMMITCOMPLETE); err != nil {
+		return fmt.Errorf("write COMMITCOMPLETE transaction info: %w", err)
+	}
+
 	wf.lastCommittedTGID = 0
 	return nil
 }
@@ -448,7 +483,7 @@ func (tgl TGIDlist) Len() int           { return len(tgl) }
 func (tgl TGIDlist) Less(i, j int) bool { return tgl[i] < tgl[j] }
 func (tgl TGIDlist) Swap(i, j int)      { tgl[i], tgl[j] = tgl[j], tgl[i] }
 
-func (wf *WALFileType) WriteStatus(fileStatus wal.FileStatusEnum, replayState wal.ReplayStateEnum) {
+func (wf *WALFileType) WriteStatus(fileStatus wal.FileStatusEnum, replayState wal.ReplayStateEnum) error {
 	wf.FileStatus = fileStatus
 	wf.ReplayState = replayState
 	// This process now owns this file
@@ -456,22 +491,30 @@ func (wf *WALFileType) WriteStatus(fileStatus wal.FileStatusEnum, replayState wa
 	buffer, _ = io.Serialize(buffer, int8(wf.FileStatus))
 	buffer, _ = io.Serialize(buffer, int8(wf.ReplayState))
 	buffer, _ = io.Serialize(buffer, wf.OwningInstanceID)
-	wf.FilePtr.Seek(0, goio.SeekStart)
-	wf.FilePtr.Write(buffer)
-	wf.FilePtr.Sync()
-	wf.FilePtr.Seek(0, goio.SeekEnd)
+	if _, err := wf.FilePtr.Seek(0, goio.SeekStart); err != nil {
+		return err
+	}
+	if _, err := wf.FilePtr.Write(buffer); err != nil {
+		return err
+	}
+	if err := wf.FilePtr.Sync(); err != nil {
+		return err
+	}
+	if _, err := wf.FilePtr.Seek(0, goio.SeekEnd); err != nil {
+		return err
+	}
+	return nil
 }
 
-func (wf *WALFileType) write(buffer []byte) {
-	wf.FilePtr.Write(buffer)
-}
-
-func (wf *WALFileType) WriteTransactionInfo(tid int64, did DestEnum, txnStatus TxnStatusEnum) {
+func (wf *WALFileType) WriteTransactionInfo(tid int64, did DestEnum, txnStatus TxnStatusEnum) error {
 	buffer := wf.initMessage(TXNINFO)
 	buffer, _ = io.Serialize(buffer, tid)
 	buffer, _ = io.Serialize(buffer, did)
 	buffer, _ = io.Serialize(buffer, txnStatus)
-	wf.write(buffer)
+	_, err := wf.FilePtr.Write(buffer)
+	if err != nil {
+	}
+	return nil
 }
 
 func (wf *WALFileType) readTransactionInfo() (tgid int64, destination DestEnum, txnStatus TxnStatusEnum, err error) {
@@ -594,7 +637,11 @@ func (wf *WALFileType) syncStatusRead() error {
 func readStatus(filePtr *os.File,
 ) (fileStatus wal.FileStatusEnum, replayStatus wal.ReplayStateEnum, owningInstanceID int64, err error) {
 	// Read from beginning of file +1 to cont over the MID
-	filePtr.Seek(1, goio.SeekStart)
+	if _, err2 := filePtr.Seek(1, goio.SeekStart); err2 != nil {
+		log.Error("failed to seek file pointer", zap.Int("offset", 1),
+			zap.String("whence", "start"), zap.Error(err2),
+		)
+	}
 	fileStatus, replayStatus, owningInstanceID, err = wal.ReadStatus(filePtr)
 	if err != nil {
 		log.Error(io.GetCallerFileContext(0) + ": Unable to ReadStatus()")
@@ -602,7 +649,11 @@ func readStatus(filePtr *os.File,
 	}
 	//	wf.FileStatus, wf.ReplayState, wf.OwningInstanceID = fileStatus, replayStatus, owningInstanceID
 	// Reset the file pointer to the end of the file
-	filePtr.Seek(0, goio.SeekEnd)
+	if _, err := filePtr.Seek(0, goio.SeekEnd); err != nil {
+		log.Error("failed to seek file pointer", zap.Int("offset", 0),
+			zap.String("whence", "end"),
+		)
+	}
 	return fileStatus, replayStatus, owningInstanceID, nil
 }
 
@@ -628,7 +679,7 @@ func (wf *WALFileType) NeedsReplay() (bool, error) {
 	return false, nil
 }
 
-func (wf *WALFileType) CanWrite(msg string, callersInstanceID int64) (bool, error) {
+func (wf *WALFileType) CanWrite(callersInstanceID int64) (bool, error) {
 	if err := wf.syncStatusRead(); err != nil {
 		return false, fmt.Errorf("read syncStatus:%w", err)
 	}
@@ -684,12 +735,18 @@ func (wf *WALFileType) SyncWAL(walRefresh, primaryRefresh time.Duration, walRota
 					}
 				}
 			case <-tickerPrimary.C:
-				wf.CreateCheckpoint()
+				if err := wf.CreateCheckpoint(); err != nil {
+					log.Error("failed to create WAL checkpoint", zap.Error(err))
+				}
 				primaryFlushCounter++
 				if primaryFlushCounter%walRotateInterval == 0 {
 					log.Info("Truncating WAL file...")
-					wf.FilePtr.Truncate(0)
-					wf.WriteStatus(wal.OPEN, wal.NOTREPLAYED)
+					if err := wf.FilePtr.Truncate(0); err != nil {
+						log.Error("failed to truncate wal file", zap.Error(err))
+					}
+					if err := wf.WriteStatus(wal.OPEN, wal.NOTREPLAYED); err != nil {
+						log.Error("failed to write NOT_REPLAYED status to wal", zap.Error(err))
+					}
 					primaryFlushCounter = 0
 				}
 			}
@@ -718,7 +775,9 @@ func (wf *WALFileType) SyncWAL(walRefresh, primaryRefresh time.Duration, walRota
 // present in the write channel, as it will flush as soon as possible.
 func (wf *WALFileType) RequestFlush() {
 	if !haveWALWriter {
-		wf.FlushToWAL()
+		if err := wf.FlushToWAL(); err != nil {
+			log.Error("failed to flush WAL", zap.Error(err))
+		}
 		return
 	}
 	// if there's already a queued flush, no need to queue another
